drm_diplay: Try to use the current connector mode
Let's try to use first the current mode of the connector instead
of trying to get the largest mode directly. This can help in cases
where the largest mode is too big. If the current mode is not valid,
we fallback to the previous behavior of selecting the largest mode.

Signed-off-by: Matheus Castello <matheus@castello.eng.br>